+++ src/main/assets/js/courts.ts
@@ -1,36 +1,17 @@
 import $ from 'jquery';
 import {orderToggleState} from '../../enums/searchToggleState';
+import {CourtsTableSearch} from './courts-table-search';
 
 export class CourtsController {
 
   private contentId = '#main-content';
-  private toggleCourts = 'toggleClosedCourtsDisplay';
-  private numberOfCourts = '#numberOfCourts';
+  private tableContainerId = '#tableContainer';
+  private toggleClosedCourtsDisplay = 'toggleClosedCourtsDisplay';
   private searchCourtsFilter = 'searchCourts';
-  private searchCourtsFilterId = '#searchCourts';
-  private courtsHiddenId = 'courtsHidden';
   private courtsNameAscToggleId = '#courtsNameAscToggle';
   private courtsUpdatedAscToggleId = '#courtsUpdatedAscToggle';
-  private courtsResults = 'courtsResults';
-  private courtsTableHeaderAsc = 'courts-table-header-asc';
-  private courtsTableHeaderDesc = 'courts-table-header-desc';
-  private courtsTableHeaderInactive = 'courts-table-header-inactive';
   private tableCourtsNameId = '#tableCourtsName';
   private tableCourtsUpdatedId = '#tableCourtsUpdated';
-  private TABLE_BODY_START = ' <tbody class="govuk-table__body">';
-  private TABLE_HEADER = ' <table class="govuk-table" id="courtsResults">' +
-    ' <thead class="govuk-table__head">' +
-    '   <tr class="govuk-table__row">' +
-    '     <th id="tableCourtsName" scope="col" ' +
-    '       class="govuk-table__header govuk-!-width-one-half courts-table-header-inactive">Name</th>' +
-    '     <th scope="col" class="govuk-table__header"></th>' +
-    '     <th id="tableCourtsUpdated" scope="col" ' +
-    '       class="govuk-table__header courts-table-header-inactive">Last Updated</th>' +
-    '     <th scope="col" class="govuk-table__header"></th>' +
-    '     <th scope="col" class="govuk-table__header"></th>' +
-    '   </tr>' +
-    ' </thead>';
-  private TABLE_HEADER_BODY_END = '</tbody> </table>';
 
   constructor() {
     this.initialize();
@@ -38,196 +19,47 @@ export class CourtsController {
 
   private initialize(): void {
     $(() => {
-      this.setUpTableData((document.getElementById(this.courtsHiddenId) as HTMLInputElement).value,
-        '', false, orderToggleState.ASC, orderToggleState.INACTIVE);
-      this.switchTableClasses($(this.courtsNameAscToggleId), $(this.tableCourtsNameId), $(this.tableCourtsUpdatedId));
-      this.setUpToggleClosedCourtsDisplay();
-      this.setUpCourtsDynamicSearchFilter();
-      this.setUpAscDecNameFilter();
-      this.setUpAscDecUpdatedDateFilter();
+      if($(this.tableContainerId).length) { // so that this doesn't load on each page
+        CourtsTableSearch.setUpTable('name', true);
+        this.setUpToggleClosedCourtsDisplay();
+        this.setUpCourtsDynamicSearchFilter();
+        this.setUpAscDecNameFilter();
+        this.setUpAscDecUpdatedDateFilter();
+      }
     });
   }
 
   private setUpToggleClosedCourtsDisplay(): void {
-    $(this.contentId).on('change', `input[name=${this.toggleCourts}]`, e => {
+    $(this.contentId).on('change', `input[name=${this.toggleClosedCourtsDisplay}]`, e => {
       e.preventDefault();
-
-      const courtsUpdatedAscToggleId = $(this.courtsUpdatedAscToggleId);
-      const courtsNameAscToggleId = $(this.courtsNameAscToggleId);
-
-      this.setUpTableData((document.getElementById(this.courtsHiddenId) as HTMLInputElement).value,
-        $(this.searchCourtsFilterId).val() as string,
-        $('#main-content input[name="toggleClosedCourtsDisplay"]').prop('checked'),
-        $(this.courtsNameAscToggleId).val() as string,
-        courtsUpdatedAscToggleId.val() as string);
-
-      courtsUpdatedAscToggleId.val() == orderToggleState.INACTIVE ?
-        this.switchTableClasses(courtsNameAscToggleId,
-          $(this.tableCourtsNameId), $(this.tableCourtsUpdatedId)) :
-        this.switchTableClasses(courtsUpdatedAscToggleId,
-          $(this.tableCourtsUpdatedId), $(this.tableCourtsNameId));
+      $(this.courtsUpdatedAscToggleId).val() == orderToggleState.INACTIVE ?
+        CourtsTableSearch.setUpTable('name') : CourtsTableSearch.setUpTable('date');
     });
   }
 
   private setUpCourtsDynamicSearchFilter(): void {
     $(this.contentId).on('input', `input[name=${this.searchCourtsFilter}]`, e => {
       e.preventDefault();
-      this.setUpTableData((document.getElementById(this.courtsHiddenId) as HTMLInputElement).value,
-        $(this.searchCourtsFilterId).val() as string,
-        $('#main-content input[name="toggleClosedCourtsDisplay"]').prop('checked'),
-        $(this.courtsNameAscToggleId).val() as string, $(this.courtsUpdatedAscToggleId).val() as string);
-      this.switchTableClasses($(this.courtsNameAscToggleId),
-        $(this.tableCourtsNameId), $(this.tableCourtsUpdatedId));
+      const updatedToggleValue = $(this.courtsUpdatedAscToggleId).val() as string;
+      (updatedToggleValue == orderToggleState.ASC || updatedToggleValue == orderToggleState.DESC) ?
+        CourtsTableSearch.setUpTable('date') : CourtsTableSearch.setUpTable('name');
     });
   }
 
   private setUpAscDecNameFilter(): void {
     $(this.contentId).on('click', `${this.tableCourtsNameId}`, e => {
       e.preventDefault();
-      this.switchTableToggle($(this.courtsNameAscToggleId), $(this.courtsUpdatedAscToggleId));
-      this.setUpTableData(
-        (document.getElementById(this.courtsHiddenId) as HTMLInputElement).value,
-        $(this.searchCourtsFilterId).val() as string,
-        $('#main-content input[name="toggleClosedCourtsDisplay"]').prop('checked'),
-        $(this.courtsNameAscToggleId).val() as string,
-        $(this.courtsUpdatedAscToggleId).val() as string);
-      this.switchTableClasses($(this.courtsNameAscToggleId),
-        $(this.tableCourtsNameId), $(this.tableCourtsUpdatedId));
+      const ascToggle = $(this.courtsNameAscToggleId);
+      CourtsTableSearch.switchTableToggle(ascToggle, $(this.courtsUpdatedAscToggleId));
+      CourtsTableSearch.setUpTable('name');
     });
   }
 
   private setUpAscDecUpdatedDateFilter(): void {
     $(this.contentId).on('click', `${this.tableCourtsUpdatedId}`, e => {
       e.preventDefault();
-      this.switchTableToggle($(this.courtsUpdatedAscToggleId), $(this.courtsNameAscToggleId));
-      this.setUpTableData(
-        (document.getElementById(this.courtsHiddenId) as HTMLInputElement).value,
-        $(this.searchCourtsFilterId).val() as string,
-        $('#main-content input[name="toggleClosedCourtsDisplay"]').prop('checked'),
-        $(this.courtsNameAscToggleId).val() as string,
-        $(this.courtsUpdatedAscToggleId).val() as string);
-      this.switchTableClasses($(this.courtsUpdatedAscToggleId),
-        $(this.tableCourtsUpdatedId), $(this.tableCourtsNameId));
+      CourtsTableSearch.switchTableToggle($(this.courtsUpdatedAscToggleId), $(this.courtsNameAscToggleId));
+      CourtsTableSearch.setUpTable('date');
     });
-  }
-
-  private setUpTableData(courts: string, searchFilterValue: string, includeClosedCourts: boolean,
-    orderNameAscendingFilter: string, orderUpdatedAscendingFilter: string): void {
-    let tableBody = this.TABLE_HEADER + this.TABLE_BODY_START;
-    const filteredCourts = this.filterCourts(courts, searchFilterValue, includeClosedCourts,
-      orderNameAscendingFilter, orderUpdatedAscendingFilter);
-    tableBody += this.getCourtsTableBody(filteredCourts) + this.TABLE_HEADER_BODY_END;
-    (document.getElementById(this.courtsResults) as HTMLElement).outerHTML = tableBody;
-    searchFilterValue.length ? $(this.numberOfCourts).show().text(
-      'Showing ' + filteredCourts.length + ' results')
-      : $(this.numberOfCourts).hide();
-  }
-
-  private filterCourts(courts: string, searchFilterValue: string, includeClosedCourts: boolean,
-    orderNameAscendingFilter: string, orderUpdatedAscendingFilter: string): any {
-
-    return JSON.parse(courts)
-      .filter((court: { displayed: boolean }) => {
-        if (includeClosedCourts) {
-          return court.displayed == true || court.displayed == false;
-        }
-        else return court.displayed == true;
-      }).filter((court: { name: string }) => {
-        if (searchFilterValue.length == 0)
-          return court;
-        else if (court.name.toLowerCase().includes(searchFilterValue.toString().toLowerCase())) {
-          return court;
-        }
-      }).sort(((a: { name: string; updated_at: string }, b: { name: string; updated_at: string }) => {
-
-        switch (orderNameAscendingFilter) {
-          case orderToggleState.ASC:
-            return (a.name > b.name) ? 1 : -1;
-          case orderToggleState.DESC:
-            return (a.name < b.name) ? 1 : -1;
-          case orderToggleState.INACTIVE:
-            break;
-        }
-
-        switch (orderUpdatedAscendingFilter) {
-          case orderToggleState.ASC:
-            return (Date.parse(a.updated_at) > Date.parse(b.updated_at)) ? 1 : -1;
-          case orderToggleState.DESC:
-            return (Date.parse(a.updated_at) < Date.parse(b.updated_at)) ? 1 : -1;
-          case orderToggleState.INACTIVE:
-            break;
-        }
-      }));
-  }
-
-  private getCourtsTableBody(filteredCourts: any): string {
-    let tableData = '';
-    $.each(filteredCourts,function(index,value) {
-      function getDataStructure(name: string, updatedAt: string, displayed: string, slug: string, frontendUrl: string): string {
-        return ' <tr class="govuk-table__row">' +
-          ' <td id = "courtTableColumnName" scope="row" class="govuk-table__cell">' + name + ' </td>' +
-          ' <td id = "courtTableColumnClosed" scope="row" class="govuk-table__cell">' + (displayed ? '' : 'closed') + ' </td>' +
-          ' <td id = "courtTableColumnLastUpdated" scope="row" class="govuk-table__cell">' + updatedAt + ' </td>' +
-          ' <td scope="row" class="govuk-table__cell">' +
-          '   <a id="view-"' + slug + ' class="govuk-link" href="' + frontendUrl +  '/courts/' + slug + '">view</a>' +
-          ' </td>' +
-          ' <td scope="row" class="govuk-table__cell">' +
-          '   <a id="edit-"' + slug + ' class="govuk-link" href="/courts/' + slug + '/edit#general">edit</a>' +
-          ' </td>' +
-          ' </tr>';
-      }
-
-      tableData += getDataStructure(value.name, value.updated_at ?? '', value.displayed, value.slug,
-        $('#courtsFrontendUrl').val() as string);
-    });
-
-    return tableData;
-  }
-
-  private switchTableToggle(toToggle: JQuery, toReset: JQuery): void {
-    const toggleUpdatedVal = toToggle.val();
-
-    switch (toggleUpdatedVal) {
-      case orderToggleState.ASC: {
-        toToggle.val(orderToggleState.DESC);
-        break;
-      }
-      case orderToggleState.DESC: {
-        toToggle.val(orderToggleState.ASC);
-        break;
-      }
-      case orderToggleState.INACTIVE: {
-        toToggle.val(orderToggleState.ASC);
-        break;
-      }
-    }
-
-    toReset.val(orderToggleState.INACTIVE);
-  }
-
-  private switchTableClasses(tableState: JQuery, thToToggle: JQuery, thToReset: JQuery): void {
-    switch (tableState.val()) {
-      case orderToggleState.ASC: {
-        thToToggle.removeClass(this.courtsTableHeaderInactive)
-          .removeClass(this.courtsTableHeaderAsc)
-          .addClass(this.courtsTableHeaderDesc);
-        break;
-      }
-      case orderToggleState.DESC: {
-        thToToggle.removeClass(this.courtsTableHeaderInactive)
-          .removeClass(this.courtsTableHeaderDesc)
-          .addClass(this.courtsTableHeaderAsc);
-        break;
-      }
-      case orderToggleState.INACTIVE: {
-        thToToggle.removeClass(this.courtsTableHeaderInactive)
-          .addClass(this.courtsTableHeaderAsc);
-        break;
-      }
-    }
-
-    thToReset.removeClass(this.courtsTableHeaderAsc)
-      .removeClass(this.courtsTableHeaderDesc)
-      .addClass(this.courtsTableHeaderInactive);
   }
 }
